Got Property Page fully working with add

+++ WebApi/Models/Lease/LeaseRequest.cs
@@ -6,6 +6,6 @@
         public DateTime EndDate { get; set; }
         public int TenantID { get; set; }
         public int PropertyID { get; set; }
-        public int MonthlyAmount { get; set; }
+        public decimal MonthlyAmount { get; set; }
     }
 }
